copy of 11.4 code

+++ sakai-11.4/samigo/samigo-api/src/java/org/sakaiproject/tool/assessment/data/ifc/questionpool/QuestionPoolDataIfc.java
@@ -126,10 +126,10 @@ public interface QuestionPoolDataIfc
   Integer getQuestionSize();
 
   // for JSF, sigh - daisyf
-  void setSubPoolSize(Long subPoolSize);
+  void setSubPoolSize(Integer subPoolSize);
 
   // for JSF, sigh - daisyf
-  Long getSubPoolSize();
+  Integer getSubPoolSize();
 
   Object clone();
 }
